+++ src/main/java/com/lewandowski/wycena3000/controller/PartController.java
@@ -4,8 +4,10 @@ import com.lewandowski.wycena3000.entity.Part;
 import com.lewandowski.wycena3000.entity.PartType;
 import com.lewandowski.wycena3000.dto.PartChangeRequestDto;
 import com.lewandowski.wycena3000.entity.Project;
+import com.lewandowski.wycena3000.security.CurrentUser;
 import com.lewandowski.wycena3000.service.PartService;
 import com.lewandowski.wycena3000.service.ProjectService;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -27,6 +29,16 @@ public class PartController {
         this.projectService = projectService;
     }
 
+    @GetMapping("/all")
+    public String all(Model model, @AuthenticationPrincipal CurrentUser currentUser) {
+        List<Part> parts = partService.getPartsByUser(currentUser.getUser());
+        Set<Long> enabledDeleteSet = partService.getEnabledDeleteSet();
+        model.addAttribute("parts", parts);
+        model.addAttribute("enabledDelete", enabledDeleteSet);
+
+        return "part/part_all";
+    }
+
     @GetMapping("/add")
     public String add(Model model) {
         Part part = new Part();
@@ -39,21 +51,26 @@ public class PartController {
     }
 
     @PostMapping("/add")
-    public String save(@Valid Part part, BindingResult result, Model model) {
+    public String save(@Valid Part part, BindingResult result, Model model, @AuthenticationPrincipal CurrentUser currentUser) {
 
         if (result.hasErrors()) {
             model.addAttribute("partTypes", partService.getPartTypes());
             return "part/part_add";
         }
 
+        part.setUser(currentUser.getUser());
         partService.save(part);
 
         return "redirect:/creator/parts/all";
     }
 
-    @GetMapping("/edit")
-    public String edit(@RequestParam long partId, Model model) {
+    @GetMapping("/edit/{partId}")
+    public String edit(@PathVariable long partId, Model model, @AuthenticationPrincipal CurrentUser currentUser) {
         Part part = partService.findById(partId);
+        if (part.getUser().getId() != currentUser.getUser().getId()) {
+            return "redirect:/creator/parts/all"; // todo add 403 screen
+        }
+
         model.addAttribute("part", part);
 
         List<PartType> partTypes = partService.getPartTypes();
@@ -63,27 +80,25 @@ public class PartController {
     }
 
     @GetMapping("/delete/{partId}")
-    public String delete(@PathVariable long partId) {
+    public String delete(@PathVariable long partId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Part part = partService.findById(partId);
+        if (part.getUser().getId() != currentUser.getUser().getId()) {
+            return "redirect:/creator/parts/all"; // todo add 403 screen
+        }
         partService.delete(partId);
 
         return "redirect:/creator/parts/all";
     }
 
-    @GetMapping("/all")
-    public String all(Model model) {
-        List<Part> parts = partService.getParts();
-        Set<Long> enabledDeleteSet = partService.getEnabledDeleteSet();
-        model.addAttribute("parts", parts);
-        model.addAttribute("enabledDelete", enabledDeleteSet);
-
-        return "part/part_all";
-    }
-
     @GetMapping("/change")
     public String changePart(@RequestParam Long partId,
-                         @RequestParam Long projectId, Model model) {
+                             @RequestParam Long projectId, Model model,
+                             @AuthenticationPrincipal CurrentUser currentUser) {
         Project project = projectService.findById(projectId);
-        List<Part> parts = partService.findAll();
+        if(project.getUser().getId() != currentUser.getUser().getId()) {
+            return "redirect:/creator/projects/all"; // todo add 403 screen
+        }
+        List<Part> parts = partService.getPartsByUser(currentUser.getUser());
 
         model.addAttribute("project", project);
         model.addAttribute("oldPartId", partId);
